requirements.txt
streamlit
Pillow
opencv-python
matplotlib
streamlit-option-menu
python-dotenv
google-generativeai